\name{coef.0.ls}
\alias{coef.0.ls}
\docType{data}
\title{
Simulation Study Parameters
}
\description{
The true parameters used in the simulation studies.
}
\usage{data("coef.0.ls")}
\format{
  The format is:
List of 3
 $ segmented :List of 32
  ..$ quadratic2b_norm       : Named num [1:5] 0 1 0 0 0
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ cubic2b_lin            : Named num [1:5] 0 1 7 0 0
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_norm          : Named num [1:5] -0.0943 0.3365 0.4005 -0.9163 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b_norm         : Named num [1:5] -0.0943 0.3365 0.4005 0 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b1_norm        : Named num [1:5] -0.0943 0.3365 0.4005 -0.2231 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b2_norm        : Named num [1:5] -0.0943 0.3365 0.4005 -0.5108 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b3_norm        : Named num [1:5] -0.0943 0.3365 0.4005 -0.0513 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b4_norm        : Named num [1:5] -0.0943 0.3365 0.4005 -0.1054 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b5_norm        : Named num [1:5] -0.0943 0.3365 0.4005 -0.1625 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b6_norm        : Named num [1:5] -0.0943 0.3365 0.4005 0.0198 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b7_norm        : Named num [1:5] -0.0943 0.3365 0.4005 0.0392 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b8_norm        : Named num [1:5] -0.0943 0.3365 0.4005 0.0583 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b9_norm        : Named num [1:5] -0.0943 0.3365 0.4005 0.9163 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b10_norm       : Named num [1:5] -0.0943 0.3365 0.4005 0 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b16_norm       : Named num [1:5] -0.0943 0.3365 0.4005 0.0198 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b17_norm       : Named num [1:5] -0.0943 0.3365 0.4005 0.0392 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b18_norm       : Named num [1:5] -0.0943 0.3365 0.4005 0.0583 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b_norm          : Named num [1:5] 0 0.336 0.4 0 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b1_norm         : Named num [1:5] 0 0.336 0.4 0.005 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b2_norm         : Named num [1:5] 0 0.336 0.4 0.01 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b3_norm         : Named num [1:5] 0 0.336 0.4 0.02 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b4_norm         : Named num [1:5] 0 0.336 0.4 0.05 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b5_norm         : Named num [1:5] 0 0.336 0.4 0.1 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_gam           : Named num [1:5] -1.3 0.336 0.4 -0.916 2.2
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_gam1          : Named num [1:5] -1 0.336 0.4 -0.916 1.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_gam2          : Named num [1:5] -0.6 0.336 0.4 -0.916 1
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2a_gam          : Named num [1:5] -0.5 0.336 0 -0.916 2.2
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ quadratic_gam          : Named num [1:5] -1.6355 0.3363 -0.0398 1.4869 2.8154
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ exp_gam                : Named num [1:5] -2.753 0.336 0.513 0.936 3.607
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ quadratic_norm_gaussian: Named num [1:5] -3.735 0.336 0.898 1.845 4.7
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ quadratic_norm         : Named num [1:5] -2.83 0.338 0.553 1.341 3.754
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ exp_norm               : Named num [1:5] -6.235 0.337 1.012 1.325 5.057
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
 $ hinge     :List of 29
  ..$ quadratic2b_norm   : Named num [1:5] 0 1 0 0 0
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ cubic2b_lin        : Named num [1:5] 0 1 7 0 0
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_norm      : Named num [1:5] -0.0943 0.3365 0.4005 -0.9163 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b_norm     : Named num [1:5] -0.0943 0.3365 0.4005 0 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b1_norm    : Named num [1:5] -0.0943 0.3365 0.4005 -0.2231 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b2_norm    : Named num [1:5] -0.0943 0.3365 0.4005 -0.5108 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b3_norm    : Named num [1:5] -0.0943 0.3365 0.4005 -0.0513 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b4_norm    : Named num [1:5] -0.0943 0.3365 0.4005 -0.1054 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b5_norm    : Named num [1:5] -0.0943 0.3365 0.4005 -0.1625 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b6_norm    : Named num [1:5] -0.0943 0.3365 0.4005 0.0198 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b7_norm    : Named num [1:5] -0.0943 0.3365 0.4005 0.0392 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b8_norm    : Named num [1:5] -0.0943 0.3365 0.4005 0.0583 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b9_norm    : Named num [1:5] -0.0943 0.3365 0.4005 0.9163 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b10_norm   : Named num [1:5] -0.0943 0.3365 0.4005 0 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b16_norm   : Named num [1:5] -0.0943 0.3365 0.4005 0.0198 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b17_norm   : Named num [1:5] -0.0943 0.3365 0.4005 0.0392 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b18_norm   : Named num [1:5] -0.0943 0.3365 0.4005 0.0583 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b_norm      : Named num [1:5] 0 0.336 0.4 0 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b1_norm     : Named num [1:5] 0 0.336 0.4 0.005 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b2_norm     : Named num [1:5] 0 0.336 0.4 0.01 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b3_norm     : Named num [1:5] 0 0.336 0.4 0.02 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b4_norm     : Named num [1:5] 0 0.336 0.4 0.05 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b5_norm     : Named num [1:5] 0 0.336 0.4 0.1 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_gam       : Named num [1:4] -0.5 0.336 -0.916 2.2
  .. ..- attr(*, "names")= chr [1:4] "(Intercept)" "z" "(x-chngpt)+" "chngpt"
  ..$ sigmoid2_gam1      : Named num [1:4] -0.2 0.336 -0.916 1.5
  .. ..- attr(*, "names")= chr [1:4] "(Intercept)" "z" "(x-chngpt)+" "chngpt"
  ..$ sigmoid2_gam2      : Named num [1:4] 0.2 0.336 -0.916 1
  .. ..- attr(*, "names")= chr [1:4] "(Intercept)" "z" "(x-chngpt)+" "chngpt"
  ..$ quadratic_gam      : Named num [1:5] -1.695 0.336 1.464 2.871 NA
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "(x-chngpt)+" "chngpt" ...
  ..$ exp_gam            : Named num [1:5] -2.046 0.334 1.044 2.152 NA
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "(x-chngpt)+" "chngpt" ...
  ..$ flatHyperbolic_norm: Named num [1:5] -4.173 0.334 3.239 4.437 NA
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "(x-chngpt)+" "chngpt" ...
 $ segmented2:List of 23
  ..$ quadratic2b_norm: Named num [1:5] 0 1 0 0 0
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ cubic2b_lin     : Named num [1:5] 0 1 7 0 0
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2_norm   : Named num [1:5] -0.0943 0.3365 0.4005 -0.9163 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b_norm  : Named num [1:5] -0.0943 0.3365 0.4005 0 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b1_norm : Named num [1:5] -0.0943 0.3365 0.4005 -0.2231 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b2_norm : Named num [1:5] -0.0943 0.3365 0.4005 -0.5108 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b3_norm : Named num [1:5] -0.0943 0.3365 0.4005 -0.0513 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b4_norm : Named num [1:5] -0.0943 0.3365 0.4005 -0.1054 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b5_norm : Named num [1:5] -0.0943 0.3365 0.4005 -0.1625 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b6_norm : Named num [1:5] -0.0943 0.3365 0.4005 0.0198 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b7_norm : Named num [1:5] -0.0943 0.3365 0.4005 0.0392 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b8_norm : Named num [1:5] -0.0943 0.3365 0.4005 0.0583 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b9_norm : Named num [1:5] -0.0943 0.3365 0.4005 0.9163 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b10_norm: Named num [1:5] -0.0943 0.3365 0.4005 0 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b16_norm: Named num [1:5] -0.0943 0.3365 0.4005 0.0198 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b17_norm: Named num [1:5] -0.0943 0.3365 0.4005 0.0392 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ sigmoid2b18_norm: Named num [1:5] -0.0943 0.3365 0.4005 0.0583 4.5
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b_norm   : Named num [1:5] 0 0.336 0.4 0 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b1_norm  : Named num [1:5] 0 0.336 0.4 0.005 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b2_norm  : Named num [1:5] 0 0.336 0.4 0.01 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b3_norm  : Named num [1:5] 0 0.336 0.4 0.02 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b4_norm  : Named num [1:5] 0 0.336 0.4 0.05 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
  ..$ smooth2b5_norm  : Named num [1:5] 0 0.336 0.4 0.1 4.8
  .. ..- attr(*, "names")= chr [1:5] "(Intercept)" "z" "x" "(x-chngpt)+" ...
}
