\name{dat.mtct.2}
\alias{dat.mtct.2}
\docType{data}
\title{
An Example Dataset
}
\description{
A dataset from the immune correlates study of Maternal To Child Transmission of HIV-1
}
\usage{dat.mtct.2}
\format{
  A data frame with 248 observations on the following 2 variables.
  \describe{
    \item{\code{NAb_score}}{a numeric vector}
    \item{\code{V3_BioV3B}}{a numeric vector}
  }
}
\references{
Permar, S. R., Fong, Y., Nathan Vandergrift, Genevieve G. Fouda, Peter Gilbert, Georgia D. Tomaras, Feng Gao and Barton F. Haynes et al.
(2015) Maternal HIV-1 Envelope variable loop 3-specific IgG responses and reduced risk of perinatal
transmission. Journal of Clinical Investigation, 125(7):2702:2706.
}
