\name{lidar}
\alias{lidar}
\docType{data}
\title{
Light Detection and Ranging Data
}
\description{
LIDAR
}
\usage{data("lidar")}
\format{
  A data frame with 221 observations on the following 2 variables.
  \describe{
    \item{\code{range}}{a numeric vector}
    \item{\code{logratio}}{a numeric vector}
  }
}
\source{
Holst, U., Hossjer, O., Bjorklund, C., Ragnarson, P. and Edner, H. (1996), Locally weighted leastsquares kernel regression and statistical evaluation of LIDAR measurements, Environmetrics,7, 401-416.
Wakefield (2013), Bayesian and Frequentist Regression Methods. Chapter 11 Spline and Kernel Methods.
}
