\name{nutrition}
\alias{nutrition}
\docType{data}
\title{
Infant Nutrition Data
}
\description{
The infant nutrition dataset comprises data collected in a study on the nutrition of infants and preschool children in the north central region of the United States of America.
}
\usage{data("nutrition")}
\format{
  A data frame with 72 observations on the following 2 variables.
  \describe{
    \item{\code{woh}}{weight/height ratio}
    \item{\code{age}}{a numeric vector}
  }
}
\source{
Eppright, E. S., Fox, H. M., Fryer, B. A., Lamkin, G. H., Vivian, V. M., Fuller, E. S. (1972). Nutrition of Infants and Preschool Children in the North Central Region of the United States of America. In World Review of Nutrition and Dietetics (Vol. 14, pp. 269-332). Karger Publishers.
}
