\name{performance.unit.test}
\alias{performance.unit.test}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Perform unit testing for performance evaluation.
}
\description{
This function performs unit testing for performance evaulation.
}
\usage{
performance.unit.test(formula.1, formula.2, family, data, B, I)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula.1}{
formula.1
}
  \item{formula.2}{
formula.2
}
  \item{family}{
family
}
  \item{data}{
data
}
  \item{B}{
B
}
  \item{I}{
I
}
}
