\name{sim.alphas}
\alias{sim.alphas}
\docType{data}
\title{
Simulation Parameters
}
\description{
Simulation Parameters
}
\usage{data(sim.alphas)}
\format{
List of 6. Names: sigmoid2_norm, sigmoid2_norm3, sigmoid3_norm, sigmoid3_norm3, sigmoid4_norm, sigmoid4_norm3. Each element is a 5x4 matrix
}
