\name{sim.hinge}
\alias{sim.hinge}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulation function
}
\description{
Simulate data for Monte Carlo study.
}
\usage{
sim.hinge(threshold.type = c("NA", "hinge"), family = c("binomial", "gaussian"),
    thres = "NA", X.ditr = "norm", mu.X, coef.X, cov.X, eps.sd, seed, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{threshold.type}{
threshold.type
}
  \item{family}{
family
}
  \item{thres}{
thres
}
  \item{X.ditr}{
X.ditr
}
  \item{mu.X}{
mu.X
}
  \item{coef.X}{
coef.X
}
  \item{cov.X}{
cov.X
}
  \item{eps.sd}{
eps.sd
}
  \item{seed}{
seed
}
  \item{n}{
n
}
}
